overlay lib: Redesign creation of the source overlay composition
The in-place upgrade itself requires to do some changes on the system to be
able to perform the in-place upgrade itself - or even to be able to evaluate
if the system is possible to upgrade. However, we do not want to (and must not)
change the original system until we pass beyond the point of not return.

For that purposes we have to create a layer above the real host file system,
where we can safely perform all operations without affecting the system
setup, rpm database, etc. Currently overlay (OVL) technology showed it is
capable to handle our requirements good enough - with some limitations.

However, the original design we used to compose overlay layer above
the host system had number of problems:
    * buggy calculation of the required free space for the upgrade RPM
      transaction
    * consumed too much space to handle partitions formatted with XFS
      without ftype attributes (even tens GBs)
    * bad UX as people had to manually adjust size of OVL disk images
    * .. and couple of additional issues derivated from problems
      listed above

The new solution prepares a disk image (represented by sparse-file)
and an overlay image for each mountpoint configured in /etc/fstab,
excluding those with FS types noted in the `OVERLAY_DO_NOT_MOUNT`
set. Such prepared OVL images are then composed together to reflect
the real host filesystem. In the end everything is cleaned.

The composition could look like this:
  orig mountpoint -> disk img -> overlay img -> new mountpoint
  -------------------------------------------------------------
    /        -> root_        -> root_/ovl        -> root_/ovl/
    /boot    -> root_boot    -> root_boot/ovl    -> root_/ovl/boot
    /var     -> root_var     -> root_var/ovl     -> root_/ovl/var
    /var/lib -> root_var_lib -> root_var_lib/ovl -> root_/ovl/var/lib
    ...

The new solution can be now problematic for system with too many partitions
and loop devices, as each disk image is loop mounted (that's same as
before, but number of disk images will be bigger in total number).
For such systems we keep for now the possibility of the fallback
to an old solution, which has number of issues mentioned above,
but it's a trade of. To fallback to the old solution, set envar:
    LEAPP_OVL_LEGACY=1

Disk images created for OVL are formatted with XFS by default. In case of
problems, it's possible to switch to Ext4 FS using:
    LEAPP_OVL_IMG_FS_EXT4=1
XFS is better optimized for our use cases (faster initialisation
consuming less space). However we have reported several issues related
to overlay images, that happened so far only on XFS filesystems.
We are not sure about root causes, but having the possibility
to switch to Ext4 seems to be wise. In case of issues, we can simple
ask users to try the switch and see if the problem is fixed or still
present.

Some additional technical details about other changes
* Added simple/naive checks whether the system has enough space on
  the partition hosting /var/lib/leapp (usually /var). Consuming the
  all space on the partition could lead to unwanted behaviour
  - in the worst case if we speak about /var partition it could mean
    problems also for other applications running on the system
* In case the container is larger than the expected min default or
  the calculation of the required free space is lower than the
  minimal protected size, return the protected size constant
  (200 MiB).

* Work just with mountpoints (paths) in the _prepare_required_mounts()
instead of with list of MountPoint named tuple. I think about the
removal of the named tuple, but let's keep it for now.

* Make apparent size of created disk images 5% smaller to protect
  failed upgrades during the transaction execution due to really
  small amount of free space.

* Cleanup the scratch directory at the end to free the consumed
  space. Disks are kept after the run of leapp when
      LEAPP_DEVEL_KEEP_DISK_IMGS=1

(cherry picked from commit d074926c75eebc56ca640e7367638bbeaa1b61a2)