Inhibit upgrade when enabled CL Rollout repositories are detected
We want to make it possible to upgrade the system
with rollout repositories present and enabled, but at the moment
the current Leapp upgrade structure makes that difficult to achieve.
Some restructuring will be required - for now, we just tell the user to
disable the rollout repos to be safe.