CLOS-3205: Rework the clearpackageconflicts actor with set definitoon to work for el8to9 upgrades
Seems like the InstalledRPM model was modified at some point, and the previous construction could not be utilized.

The logic was changed to process only package names (which is essentially already how the data was used) into the set, which is then used for package lookup.
This sidesteps the problem with the InstalledRPM type not being hashable.